+++ checkpoint/orbax/checkpoint/experimental/v1/__init__.py
@@ -21,11 +21,14 @@ Prefer to use the style::
 
 # pylint: disable=g-importing-member, g-multiple-import
 
+from orbax.checkpoint.experimental.v1 import handlers
 from orbax.checkpoint.experimental.v1._src.context import options
 from orbax.checkpoint.experimental.v1._src.context.context import (
     Context,
 )
 from orbax.checkpoint.experimental.v1._src.loading.loading import (
+    load_checkpointables,
+    load_checkpointables_async,
     load_pytree,
     load_pytree_async,
 )
@@ -33,6 +36,11 @@ from orbax.checkpoint.experimental.v1._src.metadata.loading import (
     pytree_metadata,
 )
 from orbax.checkpoint.experimental.v1._src.saving.saving import (
+    save_checkpointables,
+    save_checkpointables_async,
     save_pytree,
     save_pytree_async,
+)
+from orbax.checkpoint.experimental.v1.handlers import (
+    CheckpointableHandler,
 )